+++ src/public/assets/talk-B5Izy_OS.js
@@ -1,4 +1,4 @@
-import{c as I,r as l,j as e,C as E,h as L,l as D,L as B,M as O,a as M,B as $,H as q,T as z,o as J,n as K,p as V,q as Z}from"./card-QfmVlMIo.js";/**
+import{c as I,r as l,j as e,C as E,h as L,l as D,L as B,M as O,a as M,B as $,H as q,T as z,o as J,n as K,p as V,q as Z}from"./card-DUC_1LBB.js";/**
  * @license lucide-react v0.462.0 - ISC
  *
  * This source code is licensed under the ISC license.